Add files via upload
The "COMPLETEDEX.yaml" is a file that has every Pokemon up to gen 9. It has their typing (both nf (non fairy) and fairy typing), and their base stat totals. Different typings of Pokemon (such as galarian or primal for Kyoger/Groudon ) will have their own separate line, and will have to be specified that way in the soullink_mons.csv. The "soullink_mons.csv" is an example of how the file should be set up for the two players. You can delete the Pokemon in there to fit how you own game looks.