Add files via upload
This is the python code used to build teams. The yaml_path, use_fairy_types, soullink_csv_path, min_team_size, are all variables that need to be changed based on your commuter's file pathing, if you are in game with fairy types, and the size of team you can conceivably have.  This code is in python now, I hope to convert it into java and put it on a website in the future to help it be more accessible and run quicker.